fix: ToC challenger 2 code (#675)
Move the ToC js/css code from the experiment folder back into the default block, and also get rid of the extra challenger 1 code
