--- express/experiments/ccx0074/blocks/toc1/toc.css
@@ -1,206 +1,0 @@
-/*
- * Copyright 2022 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-main .toc-container {
-  position: relative;
-}
-
-main .toc-container .default-content-wrapper {
-  position: absolute;
-  top: 16px;
-  right: 0;
-}
-
-main .toc-container .default-content-wrapper .button-container {
-  margin-top: 0;
-}
-
-main .toc-container .default-content-wrapper .button {
-  font-size: var(--body-font-size-m);
-  font-weight: normal;
-}
-
-main .toc-container .lottie {
-  display: inline-block;
-  margin: -5px 0;
-}
-
-main .toc-container .default-content-wrapper .lottie {
-  margin-left: 0;
-  transform: scale(1.7);
-  vertical-align: bottom;
-}
-
-main .toc-container .default-content-wrapper .button:any-link,
-main .toc-container .toc-wrapper .button:any-link {
-  background: none;
-  border: none;
-  color: var(--color-info-accent);
-}
-
-main .toc-container.open .toc-wrapper {
-  display: block;
-  position: absolute;
-}
-
-main .toc-container.open .toc-wrapper:before {
-  content: '';
-  background: linear-gradient(180deg, rgba(0,0,0,.75) 0%, rgba(0,0,0,.75) 82%, rgba(64,64,64,1) 95%);
-  top: 0;
-  height: 100vh;
-  width: 100%;
-  display: block;
-  z-index: 1;
-  position: fixed;
-}
-
-main .toc-container .toc {
-  position: fixed;
-  top: 64px;
-  z-index: 2;
-  width: 100%;
-  box-sizing: border-box;
-  background-color: var(--color-gray-100);
-  background-image: url('/express/icons/cc-express-logo.svg');
-  background-position: bottom 22px left 22px;
-  background-repeat: no-repeat;
-  background-size: auto 30px;
-  font-weight: bold;
-  text-align: left;
-  max-height: 0;
-}
-
-main .toc-container.sticky .toc {
-  top: 0px;
-}
-
-.feds-header-wrapper--retracted + main .toc-container .toc {
-  top: 0px;
-}
-
-main .toc-container .toc-heading {
-  font-size: var(--body-font-size-l);
-  font-weight: 600;
-  margin-bottom: 24px;
-}
-
-main .toc-container .toc-heading:before {
-  display: inline-block;
-  content: '';
-  background: url(/express/blocks/toc/toc-sprite.svg#arrow-down) no-repeat;
-  background-size: auto 30px;
-  background-position: -156px 0;
-  height: 24px;
-  width: 18px;
-  margin-right: 12px;
-  vertical-align: middle;
-}
-
-main .toc-container .toc div div {
-  display: inline-block;
-  vertical-align: middle;
-}
-
-main .toc-container .toc .icon {
-  display: inline-block;
-  height: 18px;
-  width: 18px;
-  margin: 8px 12px 6px 0;
-  vertical-align: middle;
-}
-
-main .toc-container .toc .icon[class*='-store'] {
-  position: absolute;
-  right: 12px;
-  height: auto;
-  width: 120px;
-  margin: -8px 0 0 0;
-}
-
-main .toc-container .toc a {
-  color: var(--body-color);
-}
-
-main .toc-container .toc-close {
-  position: absolute;
-  right: 12px;
-  bottom: 8px;
-}
-
-main .toc-container .toc-close .lottie {
-  margin-left: 6px;
-  vertical-align: text-bottom;
-  transform: scale(1.7) rotateX(180deg);
-}
-
-/* Animations */
-
-main .toc-container {
-  transition: padding-top .05s linear;
-}
-
-main .toc-container .toc {
-  transition: max-height .3s ease-in-out, padding-top .3s ease, padding-bottom .3s ease, opacity .3s ease, top .3s ease;
-  max-height: 0;
-  opacity: 0;
-  overflow: hidden;
-  padding: 0 24px;
-}
-
-main .toc-container.open .toc {
-  max-height: 80vh;
-  opacity: 1;
-  padding: 24px 24px 80px;
-}
-
-main .toc-container .toc > div {
-  transition: margin-top .3s ease, opacity .3s ease .3s;
-  opacity: 0;
-}
-main .toc-container .toc > div:nth-child(2) { margin-top: -4px; }
-main .toc-container .toc > div:nth-child(3) { margin-top: -8px; }
-main .toc-container .toc > div:nth-child(4) { margin-top: -12px; }
-main .toc-container .toc > div:nth-child(5) { margin-top: -16px; }
-main .toc-container .toc > div:nth-child(6) { margin-top: -20px; }
-main .toc-container .toc > div:nth-child(7) { margin-top: -24px; }
-
-main .toc-container.open .toc > div {
-  transition: margin-top .3s ease, opacity .3s ease;
-  opacity: 1;
-}
-main .toc-container.open .toc > div:nth-child(2) { margin-top: 0px; }
-main .toc-container.open .toc > div:nth-child(3) { margin-top: 0px; }
-main .toc-container.open .toc > div:nth-child(4) { margin-top: 0px; }
-main .toc-container.open .toc > div:nth-child(5) { margin-top: 0px; }
-main .toc-container.open .toc > div:nth-child(6) { margin-top: 0px; }
-main .toc-container.open .toc > div:nth-child(7) { margin-top: 0px; }
-
-/* Side effects on rest of the page */
-
-main .toc-container.open .template-title p:first-child {
-  display: none;
-}
-
-main .toc-container.open ~ .floating-button-wrapper {
-  background: none;
-}
-
-main .floating-button .floating-button-lottie {
-  display: none;
-}
-
-main .floating-button a.button:any-link {
-  margin-right: 0;
-}
-
-main .toc-container.open ~ .floating-button-wrapper.multifunction::before {
-  display: none;
-}
